addressservice: extract out addressing logic from basichost
The goal is to keep modifying it and removing the dependency
on basichost by relying on events. At that point, we can
expose this as a separate service usable by both basic
and blank hosts
