+++ p2p/host/basic/basic_host_test.go
@@ -8,6 +8,7 @@ import (
 	"reflect"
 	"strings"
 	"sync"
+	"sync/atomic"
 	"testing"
 	"time"
 
@@ -212,10 +213,10 @@ func TestLocalIPChangesWhenListenAddrChanges(t *testing.T) {
 	h.Start()
 	defer h.Close()
 
-	h.addrMu.Lock()
-	h.filteredInterfaceAddrs = nil
-	h.allInterfaceAddrs = nil
-	h.addrMu.Unlock()
+	h.addressService.mx.Lock()
+	h.addressService.filteredInterfaceAddrs = nil
+	h.addressService.allInterfaceAddrs = nil
+	h.addressService.mx.Unlock()
 
 	// change listen addrs and verify local IP addr is not nil again
 	require.NoError(t, h.Network().Listen(ma.StringCast("/ip4/0.0.0.0/tcp/0")))
@@ -224,8 +225,8 @@ func TestLocalIPChangesWhenListenAddrChanges(t *testing.T) {
 
 	h.addrMu.RLock()
 	defer h.addrMu.RUnlock()
-	require.NotEmpty(t, h.filteredInterfaceAddrs)
-	require.NotEmpty(t, h.allInterfaceAddrs)
+	require.NotEmpty(t, h.addressService.filteredInterfaceAddrs)
+	require.NotEmpty(t, h.addressService.allInterfaceAddrs)
 }
 
 func TestAllAddrs(t *testing.T) {
@@ -619,8 +620,13 @@ func TestAddrChangeImmediatelyIfAddressNonEmpty(t *testing.T) {
 	ctx := context.Background()
 	taddrs := []ma.Multiaddr{ma.StringCast("/ip4/1.2.3.4/tcp/1234")}
 
-	starting := make(chan struct{})
+	starting := make(chan struct{}, 1)
+	var count atomic.Int32
 	h, err := NewHost(swarmt.GenSwarm(t), &HostOpts{AddrsFactory: func(addrs []ma.Multiaddr) []ma.Multiaddr {
+		// The first call here is made from the constructor. Don't block.
+		if count.Add(1) == 1 {
+			return addrs
+		}
 		<-starting
 		return taddrs
 	}})
@@ -628,11 +634,11 @@ func TestAddrChangeImmediatelyIfAddressNonEmpty(t *testing.T) {
 	defer h.Close()
 
 	sub, err := h.EventBus().Subscribe(&event.EvtLocalAddressesUpdated{})
-	close(starting)
 	if err != nil {
 		t.Error(err)
 	}
 	defer sub.Close()
+	close(starting)
 	h.Start()
 
 	expected := event.EvtLocalAddressesUpdated{
